[Order] 주문 조회 V1

+++ src/main/java/book/shop/domain/BookForm.java
@@ -3,7 +3,6 @@ package book.shop.domain;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.experimental.FieldDefaults;
-import java.math.BigDecimal;
 import static lombok.AccessLevel.PRIVATE;
 
 @Getter
@@ -12,7 +11,7 @@ import static lombok.AccessLevel.PRIVATE;
 public class BookForm {
     Long id;
     String name;
-    BigDecimal price;
+    int price;
     int stockQuantity;
     String author;
     String isbn;
