+++ src/main/java/book/shop/domain/item/Item.java
@@ -12,7 +12,6 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
 import javax.persistence.Inheritance;
 import javax.persistence.ManyToMany;
-import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 import static book.shop.enumerate.Ids.DTYPE;
@@ -33,7 +32,7 @@ public abstract class Item {
     @Column(name = ITEM_ID)
     Long id;
     String name;
-    BigDecimal price;
+    int price;
     int stockQuantity;
     @ManyToMany(mappedBy = ITEMS)
     List<Category> categories = new ArrayList<>();
